Rename example to simple-native + update mdbook + add markdownlint CI
