+++ .markdownlintrc
@@ -1,0 +1,20 @@
+ {
+      "default": false,
+      "MD001": true,
+      "MD005": true,
+      "MD006": true,
+      "MD011": true,
+      "MD014": true,
+      "MD018": true,
+      "MD019": true,
+      "MD020": true,
+      "MD021": true,
+      "MD025": true,
+      "MD027": true,
+      "MD028": true,
+      "MD031": true,
+      "MD034": true,
+      "MD037": true,
+      "MD038": true,
+      "MD013": false
+}
