+++ Cargo.toml
@@ -7,7 +7,7 @@ members = [
    "lib",
    "indexer",
    "examples/*",
-   "examples/simple-non-wasm/counter-rs",
+   "examples/simple-native/counter-rs",
    "schema"
 ]
