+++ examples/simple-native/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "simple-non-wasm"
+name = "simple-native"
 version = "0.1.0"
 edition = "2021"
