+++ examples/simple-native/contracts/counter/Forc.lock
