+++ examples/simple-native/contracts/counter/Forc.toml
