+++ examples/simple-native/contracts/counter/src/main.sw
