+++ examples/simple-native/counter-rs/Cargo.toml
