+++ examples/simple-native/frontend/index.html
