+++ examples/simple-native/manifest.yaml
